Step Definition refactoring
- One file per class
- Change base class to follow attribute naming convention
-- Change base attribute name StepDefinition
