+++ source/Xunit.Gherkin.Quick/StepDefinition/AndAttribute.cs
@@ -1,0 +1,11 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class AndAttribute : BaseStepDefinitionAttribute
+    {
+        public AndAttribute(string pattern)
+            : base("And", pattern)
+        {
+        }
+    }
+}
