+++ source/Xunit.Gherkin.Quick/StepDefinition/BaseStepDefinitionAttribute.cs
@@ -1,0 +1,32 @@
+﻿using System;
+using System.Text.RegularExpressions;
+using Xunit.Sdk;
+
+namespace Xunit.Gherkin.Quick
+{
+    [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
+    [XunitTestCaseDiscoverer("Xunit.Gherkin.Quick.IgnoreMethodDiscoverer", "Xunit.Gherkin.Quick")]
+    public abstract class BaseStepDefinitionAttribute : FactAttribute
+    {
+        public string Keyword { get; }
+
+        public string Pattern { get; }
+
+        protected BaseStepDefinitionAttribute(string keyword, string pattern)
+        {
+            Keyword = keyword ?? throw new ArgumentNullException(nameof(keyword));
+            Pattern = pattern ?? throw new ArgumentNullException(nameof(pattern));
+        }
+
+        public bool MatchesStep(string keyword, string text)
+        {
+            return Keyword.Equals(keyword.Trim(), StringComparison.OrdinalIgnoreCase)
+                && Regex.IsMatch(text.Trim(), Pattern, RegexOptions.IgnoreCase);
+        }
+
+        public Match MatchRegex(string text)
+        {
+            return Regex.Match(text.Trim(), Pattern, RegexOptions.IgnoreCase);
+        }
+    }
+}
