+++ source/Xunit.Gherkin.Quick/StepDefinition/ButAttribute.cs
@@ -1,0 +1,11 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class ButAttribute : BaseStepDefinitionAttribute
+    {
+        public ButAttribute(string pattern)
+            : base("But", pattern)
+        {
+        }
+    }
+}
