+++ source/Xunit.Gherkin.Quick/StepDefinition/GivenAttribute.cs
@@ -1,0 +1,12 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class GivenAttribute : BaseStepDefinitionAttribute
+    {
+        public GivenAttribute(string pattern)
+            : base("Given", pattern)
+        {
+
+        }
+    }
+}
