+++ source/Xunit.Gherkin.Quick/StepDefinition/ThenAttribute.cs
@@ -1,0 +1,12 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class ThenAttribute : BaseStepDefinitionAttribute
+    {
+        public ThenAttribute(string pattern)
+            : base("Then", pattern)
+        {
+
+        }
+    }
+}
