+++ source/Xunit.Gherkin.Quick/StepDefinition/WhenAttribute.cs
@@ -1,0 +1,12 @@
+﻿namespace Xunit.Gherkin.Quick
+{
+
+    public sealed class WhenAttribute : BaseStepDefinitionAttribute
+    {
+        public WhenAttribute(string pattern)
+            : base("When", pattern)
+        {
+
+        }
+    }
+}
